test: 통합 테스트명에 Controller 의미 추가 및 DB 연결 테스트 Dao 테스트로 이동
