+++ src/test/java/roomescape/AdminControllerTest.java
@@ -7,7 +7,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class AdminTest {
+public class AdminControllerTest {
 
     @Test
     void readAdmin() {
