+++ src/test/java/roomescape/ReservationTimeControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationTimeTest {
+public class ReservationTimeControllerTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
