+++ src/test/java/roomescape/dao/ReservationDaoTest.java
@@ -2,6 +2,8 @@ package roomescape.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
@@ -21,6 +23,17 @@ class ReservationDaoTest {
     private ReservationDao reservationDao;
     @Autowired
     private JdbcTemplate jdbcTemplate;
+
+    @Test
+    void getConnection() {
+        try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
+            assertThat(connection).isNotNull();
+            assertThat(connection.getCatalog()).isEqualTo("DATABASE");
+            assertThat(connection.getMetaData().getTables(null, null, "RESERVATION", null).next()).isTrue();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
     @Test
     void findAll() {
