+++ src/test/java/roomescape/dao/ReservationTimeDaoTest.java
@@ -2,6 +2,8 @@ package roomescape.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -19,6 +21,17 @@ class ReservationTimeDaoTest {
     private ReservationTimeDao reservationTimeDao;
     @Autowired
     private JdbcTemplate jdbcTemplate;
+
+    @Test
+    void getConnection() {
+        try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
+            assertThat(connection).isNotNull();
+            assertThat(connection.getCatalog()).isEqualTo("DATABASE");
+            assertThat(connection.getMetaData().getTables(null, null, "RESERVATION_TIME", null).next()).isTrue();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
     @Test
     void add() {
